requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
cffi==1.11.5
cryptography==2.4.2
fusepy==3.0.1
idna==2.7
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pycparser==2.19
pytest==4.0.1
six==1.11.0
